QR Code Generator/requirements.txt
qrcode
Pillow
ppdiffusers==0.19.3
paddlenlp==2.6.0